1. add login functon 2. delete, update articles have to own the rights

+++ JiashengBlog/article/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from django.contrib.auth.models import User
+from django.core.paginator import Paginator
+from django.contrib.auth.decorators import login_required
 from .forms import ArticlePostForm
 from .models import ArticlePost
 import markdown
@@ -23,8 +25,16 @@ def article_create(request):
 		return render(request, 'article/create.html', context)
 
 def article_list(request):
-	articles = ArticlePost.objects.all()
+	article_list = ArticlePost.objects.all()
+
+	paginator = Paginator(article_list, 4)
+
+	page = request.GET.get('page')
+
+	articles = paginator.get_page(page)
+
 	context = {'articles':articles}
+
 	return render(request, 'article/list.html', context)
 
 def article_detail(request, id):
@@ -37,11 +47,18 @@ def article_detail(request, id):
 	context = {'article':article}
 	return render(request, 'article/detail.html', context)
 
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
-	article = ArticlePost.objects.get(id=id)
-	article.delete()
-	return redirect('article:article_list')
+	user = ArticlePost.objects.get(id=id).author
 
+	if request.user == user:
+		article = ArticlePost.objects.get(id=id)
+		article.delete()
+		return redirect('article:article_list')
+	else:
+		return HttpResponse("You are not author")
+
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
 	"""
 	update the exist article by article id
@@ -49,16 +66,19 @@ def article_update(request, id):
 	the inital pages
 	"""
 	article = ArticlePost.objects.get(id=id)
-	if request.method == "POST":
-		article_post_form = ArticlePostForm(data=request.POST)
-		if article_post_form.is_valid():
-			article.title = request.POST['title']
-			article.body = request.POST['body']
-			article.save()
-			return redirect('article:article_detail', id=id)
+	if request.user == article.author:
+		if request.method == "POST":
+			article_post_form = ArticlePostForm(data=request.POST)
+			if article_post_form.is_valid():
+				article.title = request.POST['title']
+				article.body = request.POST['body']
+				article.save()
+				return redirect('article:article_detail', id=id)
+			else:
+				return HttpResponse("Erros in form, Please correct it")
 		else:
-			return HttpResponse("Erros in form, Please correct it")
+			article_post_form = ArticlePostForm()
+			context = {'article':article, 'article_post_form':article_post_form}
+			return render(request, 'article/update.html', context)
 	else:
-		article_post_form = ArticlePostForm()
-		context = {'article':article, 'article_post_form':article_post_form}
-		return render(request, 'article/update.html', context)
+		return HttpResponse("You are not author")
